+++ pages/index/index.js
@@ -1,7 +1,7 @@
 // index.js
 import Message from 'tdesign-miniprogram/message/index';
 // 获取应用实例
-const app = getApp()
+const app = getApp();
 
 Page({
   data: {
@@ -10,19 +10,21 @@ Page({
     hasUserInfo: false,
     canIUse: wx.canIUse('button.open-type.getUserInfo'),
     canIUseGetUserProfile: false,
-    canIUseOpenData: wx.canIUse('open-data.type.userAvatarUrl') && wx.canIUse('open-data.type.userNickName') // 如需尝试获取用户信息可改为false
+    canIUseOpenData:
+      wx.canIUse('open-data.type.userAvatarUrl') &&
+      wx.canIUse('open-data.type.userNickName'), // 如需尝试获取用户信息可改为false
   },
   // 事件处理函数
   bindViewTap() {
     wx.navigateTo({
-      url: '../logs/logs'
-    })
+      url: '../logs/logs',
+    });
   },
   onLoad(option) {
     if (wx.getUserProfile) {
       this.setData({
-        canIUseGetUserProfile: true
-      })
+        canIUseGetUserProfile: true,
+      });
     }
     if (option.release && option.release === 'success') {
       this.showReleaseMsg();
@@ -33,21 +35,21 @@ Page({
     wx.getUserProfile({
       desc: '展示用户信息', // 声明获取用户个人信息后的用途，后续会展示在弹窗中，请谨慎填写
       success: (res) => {
-        console.log(res)
+        console.log(res);
         this.setData({
           userInfo: res.userInfo,
-          hasUserInfo: true
-        })
-      }
-    })
+          hasUserInfo: true,
+        });
+      },
+    });
   },
   getUserInfo(e) {
     // 不推荐使用getUserInfo获取用户信息，预计自2021年4月13日起，getUserInfo将不再弹出弹窗，并直接返回匿名的用户个人信息
-    console.log(e)
+    console.log(e);
     this.setData({
       userInfo: e.detail.userInfo,
-      hasUserInfo: true
-    })
+      hasUserInfo: true,
+    });
   },
   showReleaseMsg() {
     Message.success({
@@ -57,4 +59,9 @@ Page({
       content: '发布成功',
     });
   },
-})
+  goRelease() {
+    wx.navigateTo({
+      url: '/pages/release/index',
+    });
+  },
+});
