+++ pages/index/index.less
@@ -17,3 +17,12 @@
 .usermotto {
   margin-top: 200px;
 }
+
+.home-release {
+  position: fixed;
+  bottom: 208rpx;
+  right: 32rpx;
+  .relese-button {
+    width: 200rpx;
+  }
+}
